# contracts/openapi.v1.yaml
openapi: 3.0.3
info:
  title: PM Agent API
  version: 1.0.0
  description: |
    Versioned contract between the Next.js frontend and FastAPI backend.
    Breaking changes require /v2. Authenticated routes send `Authorization: Bearer <access_token>`
    and optional `X-Workspace-Id`.
servers:
  - url: http://localhost:8000/v1
...
paths:
...
  /tickets:
    get:
      tags: [tickets]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: status, in: query, schema: { type: string } }]
      responses: { "200": { description: Tickets + activity } }
    post:
      tags: [tickets]
      security: [{ bearerAuth: [] }]
      responses: { "201": { description: Chat draft created } }
  /tickets/ingest:
    post:
      tags: [tickets]
      security: [{ bearerAuth: [] }]
      responses: { "201": { description: Ticket ingested into the workspace queue } }
  /tickets/{ticket_id}:
    get:
      tags: [tickets]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: ticket_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Ticket } }
    patch:
      tags: [tickets]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: ticket_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Updated } }
  /tickets/{ticket_id}/review:
    post:
      tags: [tickets]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: ticket_id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [accept, reject, ignore, accept_non_technical]
      responses: { "200": { description: Reviewed } }
  /activity:
    get:
      tags: [tickets]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Activity } }
  /clusters:
    get:
      tags: [tickets]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Clusters } }
  /clusters/rebuild:
    post:
      tags: [tickets]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Rebuilt } }
  /clusters/{cluster_id}:
    get:
      tags: [tickets]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: cluster_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Cluster } }
  /clusters/{cluster_id}/review:
    post:
      tags: [tickets]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: cluster_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Bulk review } }
...
  /pipeline/{card_id}:
    patch:
      tags: [workspace-ops]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: card_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Moved } }
...
  /webhooks/ingest/{workspace_id}:
    post:
      tags: [webhooks]
      parameters: [{ name: workspace_id, in: path, required: true, schema: { type: string } }]
      responses: { "201": { description: Ticket ingested } }
  /webhooks/freshdesk/{workspace_id}:
    post:
      tags: [webhooks]
      parameters: [{ name: workspace_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Freshdesk ingest } }
  /webhooks/zendesk/{workspace_id}:
    post:
      tags: [webhooks]
      parameters: [{ name: workspace_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Zendesk ingest } }
  /webhooks/jira/{workspace_id}:
    post:
      tags: [webhooks]
      parameters: [{ name: workspace_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Jira Service Management ingest } }
  /webhooks/salesforce/{workspace_id}:
    post:
      tags: [webhooks]
      parameters: [{ name: workspace_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Salesforce Case ingest } }
  /webhooks/email/{workspace_id}:
    post:
      tags: [webhooks]
      parameters: [{ name: workspace_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Inbound email ingest } }
  /webhooks/pipeline/{workspace_id}:
    post:
      tags: [webhooks]
      parameters: [{ name: workspace_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Pipeline stage update } }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...